biodata/airflow/deployment/create-service/requirements.txt
functions-framework==3.*
google-cloud-orchestration-airflow==1.*
google-api-core==2.*